refactor: extract common logic to BaseSchedulerHandler and fix import paths
